Move spec for logging to custom logger into its own file
