+++ spec/spec_helper.rb
@@ -18,5 +18,7 @@ require 'ostruct'
 require 'active_support/core_ext/string'
 require 'pry'
 require 'support'
+require 'test_doubles'
+require 'stringio'
 
 I18n.enforce_available_locales = true
